pyproject.toml
[tool.poetry]
name = "hazid"
version = "0.1.0"
description = ""
authors = ["daoolet <dauletkhabidullin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
